Update README.md
Add in a notice about the upcoming lack of intended maintenance on this repository. Sadly we haven't got the people to maintain this repository and we'd like to make it clear to visitors to this repository about it's state. This process of marking it unmaintained only affects future updates. The Action will still continue work as it is in old workflows.

Closes #78

Once this lands, we'll also Archive this repository.

If you've got good recommendations for replacements, please let me know, we'll add them into the README.